+++ plutarch-ledger-api/src/Plutarch/LedgerApi/Value.hs
@@ -56,8 +56,8 @@ module Plutarch.LedgerApi.Value (
   -- ** Combination
   pleftBiasedCurrencyUnion,
   pleftBiasedTokenUnion,
-  punionResolvingCollisionsWith,
-  punionResolvingCollisionsWithData,
+  punionWith,
+  punionWithData,
 
   -- ** Queries
   pvalueOf,
@@ -286,7 +286,7 @@ instance PEq (PValue 'AssocMap.Sorted 'NoGuarantees) where
       # (AssocMap.pall # plam (#== 0))
       -- While '(-)' is not commutative, we don't need that property here.
       -- TODO benchmark with '(==)'
-      # pto (punionResolvingCollisionsWith # plam (-) # a # b)
+      # pto (punionWith # plam (-) # a # b)
 
 -- | @since 3.3.0
 instance PSemigroup (PValue 'AssocMap.Sorted 'Positive) where
@@ -296,12 +296,12 @@ instance PSemigroup (PValue 'AssocMap.Sorted 'Positive) where
 -- | @since 2.0.0
 instance Semigroup (Term s (PValue 'AssocMap.Sorted 'Positive)) where
   a <> b =
-    punsafeDowncast (pto $ punionResolvingCollisionsWith # plam (+) # a # b)
+    punsafeDowncast (pto $ punionWith # plam (+) # a # b)
 
 -- | @since 2.0.0
 instance PlutusTx.Semigroup (Term s (PValue 'AssocMap.Sorted 'Positive)) where
   a <> b =
-    punsafeDowncast (pto $ punionResolvingCollisionsWith # plam (+) # a # b)
+    punsafeDowncast (pto $ punionWith # plam (+) # a # b)
 
 -- | @since 3.3.0
 instance PSemigroup (PValue 'AssocMap.Sorted 'NonZero) where
@@ -311,12 +311,12 @@ instance PSemigroup (PValue 'AssocMap.Sorted 'NonZero) where
 -- | @since 2.0.0
 instance Semigroup (Term s (PValue 'AssocMap.Sorted 'NonZero)) where
   a <> b =
-    pnormalize #$ punionResolvingCollisionsWith # plam (+) # a # b
+    pnormalize #$ punionWith # plam (+) # a # b
 
 -- | @since 2.0.0
 instance PlutusTx.Semigroup (Term s (PValue 'AssocMap.Sorted 'NonZero)) where
   a <> b =
-    pnormalize #$ punionResolvingCollisionsWith # plam (+) # a # b
+    pnormalize #$ punionWith # plam (+) # a # b
 
 -- | @since 3.3.0
 instance PSemigroup (PValue 'AssocMap.Sorted 'NoGuarantees) where
@@ -326,12 +326,12 @@ instance PSemigroup (PValue 'AssocMap.Sorted 'NoGuarantees) where
 -- | @since 2.0.0
 instance Semigroup (Term s (PValue 'AssocMap.Sorted 'NoGuarantees)) where
   a <> b =
-    punionResolvingCollisionsWith # plam (+) # a # b
+    punionWith # plam (+) # a # b
 
 -- | @since 2.0.0
 instance PlutusTx.Semigroup (Term s (PValue 'AssocMap.Sorted 'NoGuarantees)) where
   a <> b =
-    punionResolvingCollisionsWith # plam (+) # a # b
+    punionWith # plam (+) # a # b
 
 -- | @since 3.3.0
 instance
@@ -500,9 +500,9 @@ pcheckBinRel = phoistAcyclic $
  quantities with the same asset class. Note that the result is _not_
  'normalize'd and may contain zero quantities.
 
- @since 2.0.0
+ @since 3.5.0
 -}
-punionResolvingCollisionsWith ::
+punionWith ::
   forall (any0 :: AmountGuarantees) (any1 :: AmountGuarantees) (s :: S).
   Term
     s
@@ -511,7 +511,7 @@ punionResolvingCollisionsWith ::
         :--> PValue 'AssocMap.Sorted any1
         :--> PValue 'AssocMap.Sorted 'NoGuarantees
     )
-punionResolvingCollisionsWith = phoistAcyclic $
+punionWith = phoistAcyclic $
   plam $ \combine x y ->
     pcon . PValue $
       AssocMap.punionWith
@@ -731,9 +731,9 @@ pleftBiasedTokenUnion = phoistAcyclic $
  data-encoded quantities with the same asset class. Note that the result is
  _not_ 'normalize'd and may contain zero quantities.
 
- @since 2.1.1
+ @since 3.5.0
 -}
-punionResolvingCollisionsWithData ::
+punionWithData ::
   forall (any0 :: AmountGuarantees) (any1 :: AmountGuarantees) (s :: S).
   Term
     s
@@ -742,7 +742,7 @@ punionResolvingCollisionsWithData ::
         :--> PValue 'AssocMap.Sorted any1
         :--> PValue 'AssocMap.Sorted 'NoGuarantees
     )
-punionResolvingCollisionsWithData = phoistAcyclic $
+punionWithData = phoistAcyclic $
   plam $ \combine x y ->
     pcon . PValue $
       AssocMap.punionWith